AoC 07/A: Solution for challenge

+++ src/main/kotlin/aoc07/InputReader.kt
@@ -2,21 +2,21 @@ package aoc07
 
 fun parseEquation(line: String): Equation {
     val parts = line.split(": ")
-    val result = parts[0].toInt()
-    val operands = parts[1].split(' ').map { it.toInt() }
+    val result = parts[0].toLong()
+    val operands = parts[1].split(' ').map { it.toLong() }
     return Equation(result, operands)
 }
 
 data class Equation(
-    val result: Int,
-    val operands: List<Int>
+    val result: Long,
+    val operands: List<Long>
 )
 
 enum class Operator {
     Add, Mul
 }
 
-fun Operator.evaluate(a: Int, b: Int): Int =
+fun Operator.evaluate(a: Long, b: Long): Long =
     when (this) {
         Operator.Add -> a + b
         Operator.Mul -> a * b
